refac: move page to its own file
